DeploymentPlan::Template changes
- renamed Template#template to Template#model;
- renemed TemplateSpec to Template;
- moved blob downloading from ConfigurationHasher to Template, added
  exit status check;
- moved template binding logic from DeploymentPlanCompiler to Release
  and Template classes;
- unit tests for template binding (easier on mocks).

+++ director/lib/director/configuration_hasher.rb
@@ -4,7 +4,6 @@ module Bosh::Director
   class ConfigurationHasher
 
     class BindingHelper
-
       attr_reader :name
       attr_reader :index
       attr_reader :properties
@@ -20,55 +19,61 @@ module Bosh::Director
       def get_binding
         binding
       end
-
     end
 
+    # @param [DeploymentPlan::JobSpec]
     def initialize(job)
       @job = job
-      @templates = job.templates
       @logger = Config.logger
     end
 
-    def extract_template(job_template)
-      @template_dir = Dir.mktmpdir("template_dir")
-      temp_path = File.join(Dir.tmpdir,
-                            "template-#{UUIDTools::UUID.random_create}")
-      begin
-        File.open(temp_path, "w") do |file|
-          Config.blobstore.get(job_template.template.blobstore_id, file)
-        end
-        # TODO(lisbakke): Check tar exit code.
-        `tar -C #{@template_dir} -xzf #{temp_path}`
-      ensure
-        FileUtils.rm_f(temp_path)
+    # @param [DeploymentPlan::Template] template Template to extract
+    # @return [String] Path to a directory where template has been extracted
+    def extract_template(template)
+      temp_path = template.download_blob
+      template_dir = Dir.mktmpdir("template_dir")
+
+      output = `tar -C #{template_dir} -xzf #{temp_path} 2>&1`
+      if $?.exitstatus != 0
+        raise JobTemplateUnpackFailed,
+              "Cannot unpack `#{template.name}' job template, " +
+              "tar returned #{$?.exitstatus}, " +
+              "tar output: #{output}"
       end
+
+      template_dir
+    ensure
+      FileUtils.rm_f(temp_path) if temp_path
     end
 
+    # @param [DeploymentPlan::Template]
     def process_template(job_template)
-      extract_template(job_template)
-      manifest = YAML.load_file(File.join(@template_dir, "job.MF"))
+      template_dir = extract_template(job_template)
+      manifest = YAML.load_file(File.join(template_dir, "job.MF"))
 
-      monit_template = template_erb("monit")
+      monit_template = erb(File.join(template_dir, "monit"))
       monit_template.filename = File.join(job_template.name, "monit")
+
       templates = {}
 
       if manifest["templates"]
         manifest["templates"].each_key do |template_name|
-          template = template_erb(File.join("templates", template_name))
+          template = erb(File.join(template_dir, "templates", template_name))
           templates[template_name] = template
         end
       end
+
       @cached_templates[job_template.name] = {
         "templates" => templates,
         "monit_template" => monit_template
       }
     ensure
-      FileUtils.rm_rf(@template_dir) if @template_dir
+      FileUtils.rm_rf(template_dir) if template_dir
     end
 
     def hash
       @cached_templates = {}
-      sorted_jobs = @templates.sort { |x, y| x.name <=> y.name }
+      sorted_jobs = @job.templates.sort { |x, y| x.name <=> y.name }
       sorted_jobs.each do |job_template|
         process_template(job_template)
       end
@@ -84,12 +89,12 @@ module Bosh::Director
                                              @job.properties.to_openstruct,
                                              instance.spec.to_openstruct)
           bound_templates = bind_template(monit_template, binding_helper,
-              instance.index)
+                                          instance.index)
           templates.keys.sort.each do |template_name|
             template = templates[template_name]
             template.filename = File.join(job_template.name, template_name)
             bound_templates << bind_template(template, binding_helper,
-                instance.index)
+                                             instance.index)
             template_digest = Digest::SHA1.new
             template_digest << bound_templates
             instance_digest << bound_templates
@@ -117,8 +122,8 @@ module Bosh::Director
 
     private
 
-    def template_erb(path)
-      ERB.new(File.read(File.join(@template_dir, path)))
+    def erb(path)
+      ERB.new(File.read(path))
     end
 
   end
